// frontend/app/controller/model/workflow/commentaireDTO.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { UtilisateurDto } from "../Utilisateur.model";


export interface CommentaireDTO { 
    id?: number;
    utilisateur?: UtilisateurDto;
    stepId?: number;
    message?: string;
}